requirements/test.txt
pytest
pytest-django
pytest-pythonpath
pytest-flake8
pytest-cov
mock

# Thrid party test
django-measurement>=3.0.0
